Refactor: add DB object and splits client ui components
