+++ actions.js
@@ -1,10 +1,9 @@
-// actions.js
-// - this is where we will keep all our client actions for interacting with our
-// server
+// actions.js - this is where we will keep all our client actions for
+// interacting with our server
 var server = require('socket.io-client')('http://localhost:3003/')
 
+// create new todo item
 function create(title) {
-  // Emit the new todo as some data to the server
   server.emit('make', {
     title: title,
   })
@@ -15,6 +14,7 @@ function deleteTodo(todo) {
   server.emit('delete', todo)
 }
 
+// delete all todo items
 function deleteAll() {
   server.emit('deleteAll')
 }
@@ -24,6 +24,7 @@ function complete(todo) {
   server.emit('markComplete', todo)
 }
 
+// mark all todo items as complete
 function completeAll() {
   server.emit('completeAll')
 }
