+++ style.css
@@ -14,6 +14,7 @@ input {
 }
 
 .container {
+  width: 100%;
   max-width: 500px;
 }
 
@@ -66,8 +67,8 @@ li {
   background: none;
   border-radius: 20px;
   border: 1px solid #e4e4e4;
-  height: 20px;
-  width: 20px;
+  height: 30px;
+  width: 30px;
   margin-right: 10px;
 }
 
@@ -78,6 +79,8 @@ svg {
 
 .CheckIcon {
   margin-top: 1px;
+  stroke: #c4c4c4;
+  fill: #c4c4c4;
 }
 
 .CheckIcon.complete {
