+++ todo.js
@@ -1,3 +1,4 @@
+// todo.js - model for our todo items
 module.exports = class Todo {
   constructor(id, title = '') {
     this.id = id
